+++ distage/distage-core-api/src/main/scala/izumi/distage/model/exceptions/ConflictResolutionException.scala
@@ -1,6 +1,7 @@
 package izumi.distage.model.exceptions
 
-import izumi.distage.model.definition.errors.{ConflictResolutionError, DIError}
+import izumi.distage.model.definition.conflicts.ConflictResolutionError
+import izumi.distage.model.definition.errors.DIError
 import izumi.distage.model.plan.ExecutableOp.InstantiationOp
 import izumi.distage.model.reflection.DIKey
