+++ distage/distage-core-api/src/main/scala/izumi/distage/model/planning/ForwardingRefResolver.scala
@@ -1,6 +1,6 @@
 package izumi.distage.model.planning
 
-import izumi.distage.model.definition.errors.LoopResolutionError
+import izumi.distage.model.definition.errors.DIError.LoopResolutionError
 import izumi.distage.model.plan.ExecutableOp
 import izumi.distage.model.reflection.DIKey
 import izumi.fundamentals.graphs.DG
