+++ distage/distage-core/src/test/scala/izumi/distage/gc/GcBasicTests.scala
@@ -4,7 +4,7 @@ import distage.DIKey
 import izumi.distage.model.PlannerInput
 import izumi.distage.model.definition.errors.LoopResolutionError
 import izumi.distage.model.definition.{Activation, ModuleDef}
-import izumi.distage.model.exceptions.{InjectorFailed, UnsupportedOpException}
+import izumi.distage.model.exceptions.InjectorFailed
 import izumi.distage.model.plan.Roots
 import org.scalatest.wordspec.AnyWordSpec
