+++ distage/distage-core/src/test/scala/izumi/distage/injector/BasicTest.scala
@@ -7,7 +7,7 @@ import izumi.distage.model.PlannerInput
 import izumi.distage.model.definition.Binding.SetElementBinding
 import izumi.distage.model.definition.BindingTag
 import izumi.distage.model.definition.StandardAxis.Repo
-import izumi.distage.model.definition.errors.ConflictResolutionError
+import izumi.distage.model.definition.conflicts.ConflictResolutionError
 import izumi.distage.model.exceptions.{BadMutatorAxis, ConflictResolutionException, ProvisioningException}
 import izumi.distage.model.plan.ExecutableOp.ImportDependency
 import izumi.fundamentals.platform.functional.Identity
